GAN/Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
torch==2.1.1
torchvision==0.16.1
numpy==1.26.2
pillow==10.1.0
pyyaml==6.0.1
opencv-python==4.8.1.78
